add :split command
