+++ src/Actions/Commands/HorizontalSplit.ts
@@ -1,0 +1,8 @@
+import * as vscode from 'vscode';
+import { Command } from './Base';
+
+export class HorizontalSplitCommand extends Command {
+    execute(): Thenable<undefined> {
+        return vscode.commands.executeCommand('workbench.action.splitEditorOrthogonal');
+    }
+}
